Dynamic Input Function
Added the dynamic_input function to engine. This allows the user to pass in a vector of string options, and the user can go through the menu like in a real game.

This took so many hours to get to work. It will take some more to fix Master to be updated with the new system