Main Menu/Engine Edits
Changed the title art function so that it returns a string which enables better integration into other functions.

Removed GAME_STATE because it was only needed in one function (the new_game /load_save_game). There's now a change so that they pass a bool which dictates if its new or not.

Added string_manip, which creates a vector of strings so that they can be fed into dynamic_input with ease.

Conducted massive edits on the master.cpp file to make sure that the while loop is airtight. The next step for this is to edit the following functions
-homebase()
-start_new_game()
-load_save_file()